FEG-7 Move elements to components in Vue
