+++ docs/views/components/CardDoc.vue
@@ -73,7 +73,7 @@
 
 <script>
 import CardDeck from '@/components/CardDeck/CardDeck.vue'
-import CardReadme from '@/elements/Card/README.md'
+import CardReadme from '@/components/Card/README.md'
 
 export default {
   components: {
