+++ docs/views/components/form/Checkbox.vue
@@ -4,7 +4,7 @@
     <div class="container pb-0">
       
       <ul class="breadcrumb mb-0">
-        <li><a href="/elements/form">Form elements</a></li>
+        <li><a href="/components/form">Form components</a></li>
       </ul>
       <h1>Checkbox field</h1>
       <p>Using what a fairly common approach on styling checkboxes using a wrapper div to group the input and label.</p>
@@ -66,8 +66,8 @@
 
 
 <script>
-import Input from '@/elements/Input/Input.vue'
-import Readme from '@/elements/Input/README.md'
+import Input from '@/components/Input/Input.vue'
+import Readme from '@/components/Input/README.md'
 
 export default {
   components: {
