+++ docs/views/components/form/Textarea.vue
@@ -3,7 +3,7 @@
     
     <div class="container">
       <ul class="breadcrumb mb-0">
-        <li><a href="/elements/form">Form elements</a></li>
+        <li><a href="/components/form">Form components</a></li>
       </ul>
       <h1>Textarea</h1>
       <p>We re-use the same component as the input field and just change the input field for a textarea.</p>
@@ -25,8 +25,8 @@
 
 
 <script>
-import Input from '@/elements/Input/Input.vue'
-import Readme from '@/elements/Input/README.md'
+import Input from '@/components/Input/Input.vue'
+import Readme from '@/components/Input/README.md'
 
 export default {
   components: {
